+++ components/map/dialog/GatheringDialog.tsx
