+++ components/map/dialog/MagDialog.tsx
